pyproject.toml
[project]
name = "nutricia-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.133.1",
    "httpx>=0.28.1",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.1",
    "langchain-openai>=0.3",
    "langchain-postgres>=0.0.12",
    "pgvector>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-groq>=0.3",
    "langchain-mistralai>=0.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
